Fix using stale bind_groups when recreating pipelines
Calling Instanced::set_data can recreate the buffer if the existing one
doesn’t fit the data. This increments its generational index. The code
that responds to generation changes wasn’t recreating bind groups and
therefore the pipeline was still pointing to the old buffer that may or
may not still be around in memory.