Make worker threads wait for the main thread when decompressing
I benchmarked decompressing 662374 frames (~3 hours of footage) to see
how different channel bounds would affect the time it takes:

- bounded(0) = 7 minutes, 24 seconds
- bounded(2) = 6 minutes, 57 seconds
- bounded(10) = 5 minutes, 58 seconds

Time does decrease as the bound increases, but with bounded(10) that
would mean storing 10 x 8 threads = 80 frames in memory, which is quite
a lot of raw image data (4.5GB for 1600x1200).

In the slowest case when bounded(0), it decodes ~1500 frames/second
which is much faster than any processor will be so choose the slowest
option to save memory.