Make it so the Decompressor can take a per-thread processing closure
I want to add support for encoding to video. When sending frames to an
encoder this needs to happen in order but PngWriter messes up the order
again making this difficult.

I think a better pattern is to allow the Decompressor to take a
per-thread function that can do work (i.e. encoding to PNG) then have it
take a second function that receives frames in order along with the
output produced by the per-thread functions.