Add support for texture2DArray
In the fragment shader for my game, I have a conditional that switches
between bound textures but this is probably detrimental to performance
since the conditional needs to be evaluated for each pixel on screen.

Instead, I want to index into a texture2DArray which should(?) improve
performance since the conditional won’t be required. Therefore, add
support for this to renderer. If the number of layers for a texture is
1 then it assumes the fragment shader will use texture2D. If the number
of layers for a texture is greater than 1 then it assumes the fragment
shader will use texture2DArray.