Start working on a RenderThread struct
I want to move rendering into a background thread for my game. To
accomplish this, I plan to introduce a RenderThread struct that owns the
renderer and uses crossbeam channels for calling functions on it.

Some of the renderer functions return values, so use a separate
return_value channel with a bounded queue size of 1. When these function
calls are made, the main thread will have to block and wait for the
return value, but these calls will be made infrequently or only once
when the game loads, e.g. to create textures.

Many of the return values can’t be passed across thread boundaries
because they do not implement Send + Sync. To deal with this, introduce
a pattern of returning a ‘Ref’ which indexes into a vec owned by the
thread.